web-server/requirements.txt
fastapi[all]
fastapi_utils
python-dateutil
sse_starlette
redistimeseries
uvicorn